#347: move all output code (writer, in_memory; drivers) into own directory; adapt api docs accordingly
