+++ biocypher/output/write/_batch_writer.py
@@ -12,10 +12,12 @@ from biocypher._create import BioCypherEdge, BioCypherNode, BioCypherRelAsNode
 from biocypher._logger import logger
 from biocypher._translate import Translator
 from biocypher._deduplicate import Deduplicator
-from biocypher.write._writer import _Writer
+from biocypher.output.write._writer import _Writer
 
 
 class _BatchWriter(_Writer, ABC):
+    """Abstract batch writer class"""
+
     @abstractmethod
     def _get_default_import_call_bin_prefix(self):
         """
@@ -31,7 +33,7 @@ class _BatchWriter(_Writer, ABC):
     @abstractmethod
     def _write_array_string(self, string_list):
         """
-        Abstract method to write the string representation of an array into a .csv file.
+        Abstract method to output.write the string representation of an array into a .csv file.
         Different databases require different formats of array to optimize import speed.
 
         Args:
@@ -60,7 +62,7 @@ class _BatchWriter(_Writer, ABC):
     @abstractmethod
     def _write_edge_headers(self):
         """
-        Abstract method to write a database import-file for a graph entity that is represented
+        Abstract method to output.write a database import-file for a graph entity that is represented
         as an edge as per the definition in the `schema_config.yaml`,
         containing only the header for this type of edge.
 
@@ -306,7 +308,7 @@ class _BatchWriter(_Writer, ABC):
         """
         # TODO check represented_as
 
-        # write node data
+        # output.write node data
         passed = self._write_node_data(nodes, batch_size, force)
         if not passed:
             logger.error("Error while writing node data.")
@@ -367,9 +369,9 @@ class _BatchWriter(_Writer, ABC):
 
         else:
             # is this a problem? if the generator or list is empty, we
-            # don't write anything.
+            # don't output.write anything.
             logger.debug(
-                "No edges to write, possibly due to no matched Biolink classes.",
+                "No edges to output.write, possibly due to no matched Biolink classes.",
             )
             pass
 
@@ -523,7 +525,7 @@ class _BatchWriter(_Writer, ABC):
                         bins[label] = []
                         bin_l[label] = 0
 
-            # after generator depleted, write remainder of bins
+            # after generator depleted, output.write remainder of bins
             for label, nl in bins.items():
                 passed = self._write_single_node_list_to_file(
                     nl,
@@ -535,7 +537,7 @@ class _BatchWriter(_Writer, ABC):
                 if not passed:
                     return False
 
-            # use complete bin list to write header files
+            # use complete bin list to output.write header files
             # TODO if a node type has varying properties
             # (ie missingness), we'd need to collect all possible
             # properties in the generator pass
@@ -764,7 +766,7 @@ class _BatchWriter(_Writer, ABC):
                         bins[label] = []
                         bin_l[label] = 0
 
-            # after generator depleted, write remainder of bins
+            # after generator depleted, output.write remainder of bins
             for label, nl in bins.items():
                 passed = self._write_single_edge_list_to_file(
                     nl,
@@ -775,7 +777,7 @@ class _BatchWriter(_Writer, ABC):
                 if not passed:
                     return False
 
-            # use complete bin list to write header files
+            # use complete bin list to output.write header files
             # TODO if a edge type has varying properties
             # (ie missingness), we'd need to collect all possible
             # properties in the generator pass
@@ -960,7 +962,7 @@ class _BatchWriter(_Writer, ABC):
                 + 1
             )
 
-        # write to file
+        # output.write to file
         padded_part = str(next_part).zfill(3)
         logger.info(
             f"Writing {len(lines)} entries to {label_pascal}-part{padded_part}.csv",
@@ -993,7 +995,7 @@ class _BatchWriter(_Writer, ABC):
 
     def write_import_call(self) -> str:
         """
-        Function to write the import call detailing folder and
+        Function to output.write the import call detailing folder and
         individual node and edge headers and data files, as well as
         delimiters and database name, to the export folder as txt.
