+++ biocypher/output/write/_get_writer.py
@@ -14,13 +14,13 @@ suitable for import into a DBMS.
 """
 
 from biocypher._logger import logger
-from biocypher.write.graph._rdf import _RDFWriter
-from biocypher.write.graph._neo4j import _Neo4jBatchWriter
-from biocypher.write.graph._arangodb import _ArangoDBBatchWriter
-from biocypher.write.graph._networkx import _NetworkXWriter
-from biocypher.write.relational._csv import _PandasCSVWriter
-from biocypher.write.relational._sqlite import _SQLiteBatchWriter
-from biocypher.write.relational._postgresql import _PostgreSQLBatchWriter
+from biocypher.output.write.graph._rdf import _RDFWriter
+from biocypher.output.write.graph._neo4j import _Neo4jBatchWriter
+from biocypher.output.write.graph._arangodb import _ArangoDBBatchWriter
+from biocypher.output.write.graph._networkx import _NetworkXWriter
+from biocypher.output.write.relational._csv import _PandasCSVWriter
+from biocypher.output.write.relational._sqlite import _SQLiteBatchWriter
+from biocypher.output.write.relational._postgresql import _PostgreSQLBatchWriter
 
 logger.debug(f"Loading module {__name__}.")
 
@@ -69,19 +69,14 @@ def get_writer(
     file.
 
     Args:
-
         dbms: the database management system; for options, see DBMS_TO_CLASS.
-
         translator: the Translator object.
-
-        output_directory: the directory to write the output files to.
-
+        deduplicator: the Deduplicator object.
+        output_directory: the directory to output.write the output files to.
         strict_mode: whether to use strict mode.
 
     Returns:
-
         instance: an instance of the selected writer class.
-
     """
 
     dbms_config = _config(dbms)
