+++ biocypher/output/write/graph/_rdf.py
@@ -24,7 +24,7 @@ from rdflib.namespace import (
 
 from biocypher._create import BioCypherEdge, BioCypherNode
 from biocypher._logger import logger
-from biocypher.write._batch_writer import _BatchWriter
+from biocypher.output.write._batch_writer import _BatchWriter
 
 
 class _RDFWriter(_BatchWriter):
@@ -169,6 +169,7 @@ class _RDFWriter(_BatchWriter):
 
             # add properties to the transformed edge --> node
             for key, value in rdf_properties.items():
+                # only write value if it exists.
                 # only write value if it exists.
                 if value:
                     self.add_property_to_graph(graph, rdf_predicate, value, key)
