+++ docs/output/index.rst
@@ -11,19 +11,13 @@ and will update the documentation accordingly.
 
 The used output format is specified via the ``dbms`` parameter in the
 ``biocypher_config.yaml`` (see the :ref:`config` for an example).  Currently
-supported are ``neo4j``, ``arangodb``, ``rdf``, ``postgres``, and ``sqlite``.
+supported are ``neo4j``, ``arangodb``, ``rdf``, ``csv``, ``postgres``, ``sqlite``, and ``networkx``.
 
 Furthermore, you can specify whether to use the ``offline`` or ``online`` mode.
 
-- For the online mode set ``offline: false``. You need a running database
-instance and BioCypher will connect to this instance and directly writes the
-output to the database.
+- For the online mode set ``offline: false``. You need a running database instance and BioCypher will connect to this instance and directly writes the output to the database.
 
-- For the offline mode set ``offline: true``. BioCypher will write the knowledge
-graph to files in a designated output folder (standard being ``biocypher-out/``
-and the current datetime). Furthermore you can generate a bash script to insert
-the knowledge graph files into the specified ``dbms`` by running
-``bc.write_import_call()``.
+- For the offline mode set ``offline: true``. BioCypher will output.write the knowledge graph to files in a designated output folder (standard being ``biocypher-out/`` and the current datetime). Furthermore you can generate a bash script to insert the knowledge graph files into the specified ``dbms`` by running ``bc.write_import_call()``.
 
 .. caution::
 
@@ -37,6 +31,8 @@ supported output formats are described on the following pages:
 
     neo4j.md
     arangodb.md
+    csv.md
     rdf.md
     postgres.md
     sqlite.md
+    networkx.md
